refactor: split db/index.ts into focused modules, update CLAUDE.md
db/index.ts (899 lines) split into:
- db/types.ts — all TypeScript interfaces
- db/schema.ts — DDL, migrations, connection management
- db/chunking.ts — chunkText, sanitizeFtsQuery
- db/queries.ts — core CRUD operations
- db/analytics.ts — aggregation/reporting queries
- db/index.ts — thin re-export barrel (all callers unchanged)

CLAUDE.md updated to reflect current architecture: new src/log.ts,
format.ts, tools.ts, install/, learn/, profiles/ modules; CLI commands
section added; phase table extended through phase 9.
